backend/requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==3.7.1
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
fastapi==0.104.1
google-ai-generativelanguage==0.4.0
google-api-core==2.25.1
google-auth==2.40.3
google-generativeai==0.3.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.10
joblib==1.5.2
lxml==6.0.1
neo4j==5.15.0
numpy==1.26.2
pandas==2.1.4
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.5.0
pydantic_core==2.14.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2025.2
requests==2.32.5
rsa==4.9.1
scikit-learn==1.3.2
scipy==1.16.1
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.8.0
tzdata==2025.2
urllib3==2.5.0
